json voteAverge happen int

+++ lib/src/model/movie_now_playing_info.dart
@@ -43,7 +43,7 @@ class MovieNowPlayingInfo {
     releaseDate = json['release_date'];
     title = json['title'];
     video = json['video'];
-    voteAverage = json['vote_average'];
+    voteAverage = json['vote_average'].toDouble() ;
     voteCount = json['vote_count'];
   }
